refactor(error_tracking): move external references to presentation
