+++ products/error_tracking/backend/presentation/__init__.py
@@ -1,0 +1,8 @@
+from .serializers import ErrorTrackingExternalReferenceIntegrationSerializer, ErrorTrackingExternalReferenceSerializer
+from .views import ErrorTrackingExternalReferenceViewSet
+
+__all__ = [
+    "ErrorTrackingExternalReferenceIntegrationSerializer",
+    "ErrorTrackingExternalReferenceSerializer",
+    "ErrorTrackingExternalReferenceViewSet",
+]
